requirements.txt
click
psycopg2-binary
influxdb
python-dotenv
autopep8